Abstract methods to their own files
