+++ megaphone-validator/index.js
@@ -1,63 +1,13 @@
-const mm = require('music-metadata');
+const getAudioDuration = require('./lib/get-audio-duration');
 const request = require('request');
+const request = require('./lib/reupload-audio');
+const searchForMegaphoneEpisodes = require('./lib/search-megaphone');
 
 // Load environment variables
 require('dotenv').config();
 
-const authorizationHeader = {
-    Authorization: `Token token=${process.env.MEGAPHONE_AUTH_TOKEN}`
-}
 const programSlugString = process.env.SCPR_PROGRAM_SLUGS || '';
 const differenceThreshold = parseInt(process.env.DIFFERENCE_THRESHOLD) || 6;
-
-const searchForMegaphoneEpisodes = (externalId) => {
-    return new Promise((resolve, reject) => {
-        const options = {
-            headers: authorizationHeader,
-            json: true,
-            method: 'GET',
-            url: `https://cms.megaphone.fm/api/search/episodes?externalId=${externalId}`,
-        }
-
-        request(options, (error, response, body) => {
-            if (error) reject(error)
-            else if (body && body.length > 0) {
-                const episode = body[0];
-                resolve(episode);
-            }
-        });
-    });
-}
-
-const getAudioDuration = (url) => {
-    return new Promise((resolve, reject) => {
-        request({ url, encoding: null }, (err, response, body) => {
-            mm.parseBuffer(body, 'audio/mpeg', { duration: true })
-                .then(metadata => resolve(metadata.format.duration))
-                .catch(err => reject(err));
-        });
-    });
-}
-
-const reuploadAudio = (megaphoneEpisode, audioUrl) => {
-    const episodeId = megaphoneEpisode.id;
-    const networkId = process.env.MEGAPHONE_NETWORK_ID;
-    const podcastId = megaphoneEpisode.podcastId;
-    const options = {
-        body: {
-            backgroundAudioFileUrl: audioUrl
-        },
-        headers: authorizationHeader,
-        json: true,
-        method: 'PUT',
-        url: `https://cms.megaphone.fm/api/networks/${networkId}/podcasts/${podcastId}/episodes/${episodeId}`,
-    };
-
-    request(options, (err, response, body) => {
-        console.log(body);
-        console.log(body.audioFileProcessing);
-    });
-}
 
 exports.handler = (event, context, callback) => {
     // Exit if there are no podcasts indicated in the env
