+++ megaphone-validator/lib/get-audio-duration.js
@@ -1,0 +1,12 @@
+const mm = require('music-metadata');
+const request = require('request');
+
+module.exports = (url) => {
+    return new Promise((resolve, reject) => {
+        request({ url, encoding: null }, (err, response, body) => {
+            mm.parseBuffer(body, 'audio/mpeg', { duration: true })
+                .then(metadata => resolve(metadata.format.duration))
+                .catch(err => reject(err));
+        });
+    });
+};
